Hindu-Jeevan-Backend/app/auth.py
from passlib.context import CryptContext
from jose import jwt
from datetime import datetime, timedelta
from typing import Dict

# Password hashing configuration
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# JWT configuration — make sure to replace with a secure key in production
SECRET_KEY = "super-secret"  # 🔐 Replace with os.environ.get("SECRET_KEY") in production
ALGORITHM = "HS256"


def hash_password(password: str) -> str:
    return pwd_context.hash(password)


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)


def create_access_token(data: Dict, expires_minutes: int = 60) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=expires_minutes)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
