Hindu-Jeevan-Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
requests
google-api-python-client
oauth2client
